Add files via upload
The goal of this project is to create a profile page where users can perform CRUD operations -
Create, Read, Update, and Delete - on their profiles. The system will be built using HTML, CSS,
PHP, and MySQL technologies. This project is intended to assess the candidate's understanding
and proficiency in web development fundamentals.